[TOkenizer] Seprate Tokenizer
